Inventory :: changed from long to double

+++ dpdocter/src/main/java/com/dpdocter/beans/Drug.java
@@ -47,7 +47,7 @@ public class Drug extends GenericCollection {
 
 	private Long totalStock;
 
-	private Long retailPrice;
+	private Double retailPrice;
 
 	public String getId() {
 		return id;
@@ -201,11 +201,11 @@ public class Drug extends GenericCollection {
 		this.totalStock = totalStock;
 	}
 
-	public Long getRetailPrice() {
+	public Double getRetailPrice() {
 		return retailPrice;
 	}
 
-	public void setRetailPrice(Long retailPrice) {
+	public void setRetailPrice(Double retailPrice) {
 		this.retailPrice = retailPrice;
 	}
